Completed but untested Bank Model

+++ BankConsoleApp/Simulator.java
@@ -6,6 +6,7 @@ import java.util.Iterator;
 import java.util.Map;
 //import java.util.Random;
 import java.util.Scanner;
+import java.util.Set;
 
 public class Simulator {
 	private static ArrayList<Client> clientDB = new ArrayList<Client>();
@@ -22,15 +23,17 @@ public class Simulator {
 				+ "Add balance - add [CLIENT NAME] [ACCOUNT ID] [BALANCE]\n"
 				+ "Check client - check [CLIENT NAME]\n"
 				+ "Remove client - remove [CLIENT NAME]\n"
-				+ "Remove account - remove [CLIENT NAME] account\n"
+				+ "Remove account - remove [CLIENT NAME] [ACCOUNT ID]\n"
 				+ "Remove balance - remove [CLIENT NAME] [ACCOUNT ID] [BALANCE]\n"
 				+ "Transfer funds - transfer [SOURCE ACCOUNT ID] [DESTINATION ACCOUNT ID] [BALANCE]\n"
 				+ "Quit - quit\n\n"
 				+ "Your input here: ");
 	}
 	
-	private static void addClient(Client client) {
-		clientDB.add(client);
+	private static void addClient(String newClient) {
+		Client c = new Client(newClient, clientID);
+		clientDB.add(c);
+		clientID++;
 	}
 	
 	private static void checkClient(String checkedClient) {
@@ -50,8 +53,12 @@ public class Simulator {
 		}
 	}
 	
-	private static void removeClient(Client client) {
-		clientDB.remove(client);
+	private static void removeClient(String rmvedClient) {
+		for (Client client : clientDB) {
+			if (client.getName().equals(rmvedClient)) {
+				clientDB.remove(client);
+			}
+		}
 	}
 	
 	private static String[] collectInput() { 
@@ -63,9 +70,34 @@ public class Simulator {
 		String instruction = command[0].toLowerCase();
 		switch (instruction) {
 			case "add":
+				switch (command.length) {
+					case 2:
+						addClient(command[1]);
+						break;
+						
+					case 3:
+						addAccount(command[1], accountID);
+						accountID++;
+						
+					case 4:
+						addBalance(command[1], Integer.parseInt(command[2]), 
+								Integer.parseInt(command[3]));
+				}
 				break;
 				
 			case "remove":
+				switch (command.length) {
+					case 2:
+						removeClient(command[1]);
+						break;
+					case 3:
+						removeAccount(command[1], Integer.parseInt(command[2]));
+						break;
+					case 4:
+						removeBalance(command[1], Integer.parseInt(command[2]), 
+								Integer.parseInt(command[3]));
+						break;
+				}
 				break;
 				
 			case "check":
@@ -87,6 +119,50 @@ public class Simulator {
 			default:
 				System.out.println("I'm sorry, what was that?");
 				break;
+		}
+	}
+	
+	private static void addAccount(String clientName, int accountID) {
+		for (Client client : clientDB) {
+			if (client.getName().equals(clientName)) {
+				client.addAccount(accountID, 0);
+			}
+		}
+	}
+	
+	private static void removeAccount(String clientName, int accountID) {
+		for (Client client : clientDB) {
+			if (client.getName().equals(clientName)) {
+				client.removeAccount(accountID);
+			}
+		}
+	}
+	
+	private static void addBalance(String clientName, int accountID, int balance) {
+		for (Client client : clientDB) {
+			if (client.getName().equals(clientName)) {
+				HashMap<Integer, Integer> clientAccts = client.getAccounts();
+				Set<Integer> acctSet = clientAccts.keySet(); 
+				for (int id : acctSet) {
+					if (id == accountID) {
+						client.addAmount(accountID, balance);
+					}
+				}
+			}
+		}
+	}
+	
+	private static void removeBalance(String clientName, int accountID, int balance) {
+		for (Client client : clientDB) {
+			if (client.getName().equals(clientName)) {
+				HashMap<Integer, Integer> clientAccts = client.getAccounts();
+				Set<Integer> acctSet = clientAccts.keySet(); 
+				for (int id : acctSet) {
+					if (id == accountID) {
+						client.removeAmount(accountID, balance);
+					}
+				}
+			}
 		}
 	}
